refactor(cli): restructure commands into keeweb/ and cliproxy/ subdirectories
Move flat <app>-<action>.ts layout to <app>/<action>.ts with barrel
index.ts per group. cli.ts now imports 2 barrels + mcp standalone.

- Move 8 source files + 7 test files into subdirectories
- Create keeweb/index.ts and cliproxy/index.ts barrels
- Fix all import.meta.dir relative paths (+1 depth level)
- Fix all cross-file imports (test files, dynamic imports)
- Fix test cliDir resolution (../.. → ../../..)
- Add implementation plan at docs/plans/

All 104 tests pass, tsc clean, 0 lint errors.
