+++ packages/cli/src/cli.ts
@@ -2,26 +2,16 @@
 
 import {goke} from 'goke'
 import pkg from '../package.json' with {type: 'json'}
-import {registerCliproxyConfig} from './commands/cliproxy-config'
-import {registerCliproxyDeploy} from './commands/cliproxy-deploy'
-import {registerCliproxyKeys} from './commands/cliproxy-keys'
-import {registerCliproxyLogin} from './commands/cliproxy-login'
-import {registerCliproxyStatus} from './commands/cliproxy-status'
-import {registerKeewebDeploy} from './commands/keeweb-deploy'
-import {registerKeewebStatus} from './commands/keeweb-status'
+import {registerCliproxyCommands} from './commands/cliproxy'
+import {registerKeewebCommands} from './commands/keeweb'
 import {registerMcp} from './commands/mcp'
 
 const cli = goke('infra')
 
 cli.option('--verbose', 'Enable verbose output for all commands')
 
-registerKeewebStatus(cli)
-registerKeewebDeploy(cli)
-registerCliproxyStatus(cli)
-registerCliproxyDeploy(cli)
-registerCliproxyConfig(cli)
-registerCliproxyKeys(cli)
-registerCliproxyLogin(cli)
+registerKeewebCommands(cli)
+registerCliproxyCommands(cli)
 registerMcp(cli)
 
 cli.help()
