+++ packages/cli/src/commands/cliproxy/deploy.ts
@@ -11,12 +11,12 @@ const WORKFLOW_URL = 'https://github.com/marcusrbrown/infra/actions/workflows/de
 type CliInstance = ReturnType<typeof goke>
 
 export function resolveLocalDeployScriptPath(): string {
-  const primary = resolve(import.meta.dir, '../../../../apps/cliproxy/src/deploy.ts')
+  const primary = resolve(import.meta.dir, '../../../../../apps/cliproxy/src/deploy.ts')
   if (existsSync(primary)) {
     return primary
   }
 
-  return resolve(import.meta.dir, '../../../apps/cliproxy/src/deploy.ts')
+  return resolve(import.meta.dir, '../../../../apps/cliproxy/src/deploy.ts')
 }
 
 export function getLocalDeployEnv(): Record<string, string> {
