+++ packages/cli/src/commands/cliproxy/index.ts
@@ -1,0 +1,15 @@
+import type {goke} from 'goke'
+
+import {registerCliproxyConfig} from './config'
+import {registerCliproxyDeploy} from './deploy'
+import {registerCliproxyKeys} from './keys'
+import {registerCliproxyLogin} from './login'
+import {registerCliproxyStatus} from './status'
+
+export function registerCliproxyCommands(cli: ReturnType<typeof goke>): void {
+  registerCliproxyStatus(cli)
+  registerCliproxyDeploy(cli)
+  registerCliproxyConfig(cli)
+  registerCliproxyKeys(cli)
+  registerCliproxyLogin(cli)
+}
