+++ packages/cli/src/commands/cliproxy/status.ts
