+++ packages/cli/src/commands/keeweb/index.ts
@@ -1,0 +1,9 @@
+import type {goke} from 'goke'
+
+import {registerKeewebDeploy} from './deploy'
+import {registerKeewebStatus} from './status'
+
+export function registerKeewebCommands(cli: ReturnType<typeof goke>): void {
+  registerKeewebStatus(cli)
+  registerKeewebDeploy(cli)
+}
